Caps Lock Support
I added full support for the Caps Lock.  The firware can now check with the OS to see if the Caps Lock is active.  The firmware no longer has to keep track of it's own Caps Lock state.  I tested this by pressing the caps lock on another keyboard.  The Pico 87 recognizes this and toggles it's own LED_1.  I also increased both RepeatDelay ints, as it was impossin;e to play games with the values it was at.  Most OSes repeat keys that are held down at their own rate anyway.